+++ src/cjs/utils/helpers/VEllipseSectorHelper.d.ts
@@ -4,7 +4,7 @@
  * @author   Ikaros Kappler
  * @date     2021-03-24
  * @modified 2025-04-02 Adding `VEllipseSectorHelper.drawHandleLines`.
- * @modified 2025-05-07 Modifying the calculation of `startAngle` and `endAngle` from the rotation control point: wrapping result into [0,TWO_PI).
+ * @modified 2025-04-07 Modifying the calculation of `startAngle` and `endAngle` from the rotation control point: wrapping result into [0,TWO_PI).
  * @version  1.1.0
  */
 import { VEllipseSector } from "../../VEllipseSector";
@@ -15,7 +15,27 @@ export declare class VEllipseSectorHelper {
     private startAngleControlPoint;
     private endAngleControlPoint;
     private rotationControlPoint;
+    private _rotationControlLine;
+    private _startAngleControlLine;
+    private _endAngleControlLine;
+    private _centerHandler;
+    private _rotationHandler;
+    private _startAngleHandler;
+    private _endAngleHandler;
     constructor(sector: VEllipseSector, startAngleControlPoint: Vertex, endAngleControlPoint: Vertex, rotationControlPoint: Vertex);
+    /**
+     * Creates a new drag handler for the circle sector's start control point.
+     *
+     * @private
+     * @method _handleDragStartControlPoint
+     * @instance
+     * @memberof CircleSectorHelper
+     * @returns A new event handler.
+     */
+    private _handleDragCenterPoint;
+    private _handleDragRotationControlPoint;
+    private _handleDragStartAngleControlPoint;
+    private _handleDragEndAngleControlPoint;
     /**
      * Draw grey handle lines.
      *
@@ -23,4 +43,8 @@ export declare class VEllipseSectorHelper {
      * @param {DrawLib<any>} fill - The fill library instance to use.
      */
     drawHandleLines(draw: DrawLib<any>, fill: DrawLib<any>): void;
+    /**
+     * Destroy this VEllipseHandler which means: all listeners are being removed.
+     */
+    destroy(): void;
 }